fix: close vacuous reaper test, stale falsification claim, flake (P4)
Moves the "reaper resumes after a snapshot" check in
MemberExpiryReaperSkipsContainerDuringConcurrentSnapshot to before the
debug-reload round trip: run after reload, rdb_load.cc's own lazy-expiry
check had already dropped the due fields regardless of the reaper, so the
assertion held whether or not DeleteExpiredStep did anything. Falsified in
its new position.

Rewrites MemberExpiryReaperDoesNotBlockOnConcurrentBgsave's comment, which
still claimed reverting the reaper's delete mechanism aborts the process via
an atomic-section check -- unreproducible now that HasRegisteredCallbacks()
skips the reap before it reaches any delete mechanism. States plainly that
the round-1 redirect to the three-arg Del no longer has its own regression
test as a result (defense-in-depth only).

Closes a flake window in both snapshot-race tests: assertions checking "the
reaper did not touch X" ran after the snapshot consumer unregistered, so the
real background heartbeat could reap first. Pins the walk budget to 0 by
default, restoring it only for the synchronous windows that want the reaper
active.

Also: casts oah_table.cc's capacity_log_ assignment directly to its own
uint8_t type instead of through uint32_t, and drops a stale line-number
citation in docs/PLAN.md.

+++ src/core/oah_table.cc
@@ -80,7 +80,10 @@ template <typename Entry> size_t OAHTable<Entry>::SizeSlow() {
 template <typename Entry> void OAHTable<Entry>::GrowCapacity(size_t bucket_capacity) {
   bucket_capacity = std::max(kMinBucketCount, absl::bit_ceil(bucket_capacity));
   if (bucket_capacity > entries_.size()) {
-    capacity_log_ = uint32_t(absl::bit_width(bucket_capacity) - 1);
+    // drakeydb: P4-0 Task 2b, fix round 8 (minor) -- cast directly to capacity_log_'s own
+    // type (uint8_t since fix round 6's OAHSet sizeof regroup), not to uint32_t first: the
+    // old cast read as if the field were still 32-bit, which it hasn't been since that fix.
+    capacity_log_ = static_cast<uint8_t>(absl::bit_width(bucket_capacity) - 1);
     size_t prev_size = entries_.size();
     entries_.resize(Capacity());
     Rehash(prev_size);
